PGN headers show on analyzed AI game

+++ chonse2/src/app/homepage/homepage.ts
@@ -10,6 +10,7 @@ import GameLinkHelper from '../chessboard/chessboard/game-link-helper';
 import { ToastrService } from 'ngx-toastr';
 import { GameState } from '../../lib/game-state';
 import { RouteConstants } from '../app.routes';
+import { PgnHeaders } from '../chessboard/chessboard/pgn-misc';
 
 @Component({
   selector: 'app-homepage',
@@ -27,6 +28,7 @@ export class Homepage implements OnInit{
   vsAiStates: Array<Chonse2> | undefined;
   vsAiGameStates: Array<GameState> | undefined;
   vsAiMoves: Array<IMoveResult> | undefined;
+  vsAiPgnHeaders: PgnHeaders | undefined;
 
   progress: number = 0;
   setProgress = (n: number) =>
@@ -55,6 +57,7 @@ export class Homepage implements OnInit{
     this.vsAiStates = state[RouteConstants.ROUTE_VSAI_STATES];
     this.vsAiGameStates = state[RouteConstants.ROUTE_VSAI_GAMESTATES];
     this.vsAiMoves = state[RouteConstants.ROUTE_VSAI_MOVES];
+    this.vsAiPgnHeaders = state[RouteConstants.ROUTE_VSAI_PGNHEADERS];
 
     if (this.site && this.username && this.gameId)
     {
@@ -104,20 +107,28 @@ export class Homepage implements OnInit{
       this.gameService.deleteGame(BoardNames.Analysis);
       this.gameService.addGame(BoardNames.Analysis, boardState);
     }
-    else if (this.vsAiMoves && this.vsAiStates && this.vsAiGameStates)
+    else if (this.vsAiMoves && this.vsAiStates && this.vsAiGameStates && this.vsAiPgnHeaders)
     {
+      //Set up board state.
       const bs = new BoardState();
       bs.isReadOnly = true;
       bs.doEvaluateGame = true;
       
+      //Set positions.
       const restoredPositions = this.vsAiStates.map( s => Object.assign(new Chonse2, s) );
       bs.mainStateStack = restoredPositions;
 
+      //Set game states for the positions.
       const restoredGameStates = this.vsAiGameStates?.map( s => Object.assign(new GameState, s) );
       restoredPositions.forEach( (s: Chonse2, idx: number) => s.gameState = restoredGameStates[idx]);
-
+  
+      //Set move stack.
       bs.mainMoveStack = this.vsAiMoves;
 
+      //Set pgn headers.
+      bs.pgnHeaders = this.vsAiPgnHeaders;
+
+      //Add game to service.
       this.gameService.deleteGame(BoardNames.Analysis);
       this.gameService.addGame(BoardNames.Analysis, bs);
     }
